requirements.txt
deepl
python-dotenv
PyMuPDF
deep-translator